+++ segments_datas.txt
@@ -1,0 +1,1 @@
+segments_datas = [[[(1651.5351064527194, -290.74132573255923), (1646.5341064527197, -290.74132573255923)], [(1646.5341064527197, -290.74132573255923), (1646.5341064527197, -303.5872622138172)], [(1651.5271590241482, -290.738027446845), (1645.7771590241482, -290.738027446845)], [(1645.7771590241482, -290.738027446845), (1634.2771590241482, -291.19703965638445)], [(1634.2771590241482, -291.19703965638445), (1633.5271590241482, -291.2195396563844)], [(1633.5271590241482, -291.2195396563844), (1624.5271590241482, -297.2195396563844)], [(1624.5271590241482, -297.2195396563844), (1623.5271590241482, -297.2495396563844)], [(1623.5271590241482, -297.2495396563844), (1616.0861103938246, -301.50156744514067)], [(1651.5271590241482, -298.58735144684476), (1646.517159024148, -298.58735144684476)], [(1646.517159024148, -298.58735144684476), (1646.1591590241483, -298.57635144684474)], [(1646.1591590241483, -298.57635144684474), (1639.276159024148, -298.65935144684477)], [(1639.276159024148, -298.65935144684477), (1635.8651590241484, -298.7333514468447)], [(1635.8651590241484, -298.7333514468447), (1635.7571590241482, -298.74135144684476)], [(1635.7571590241482, -298.74135144684476), (1633.082159024148, -298.9423514468447)], [(1633.082159024148, -298.9423514468447), (1626.9311590241482, -299.0773514468448)], [(1626.9311590241482, -299.0773514468448), (1626.9271590241483, -299.0773514468448)], [(1626.9271590241483, -299.0773514468448), (1626.1241590241484, -299.07035144684477)], [(1626.1241590241484, -299.07035144684477), (1621.2101590241482, -299.02835144684474)], [(1621.2101590241482, -299.02835144684474), (1621.1041590241484, -299.02735144684476)], [(1621.1041590241484, -299.02735144684476), (1615.4275551501166, -298.9730145854379)]]]
